fix: remember prev selected lang when creating new code block
Previously, on creating a new code block, Javascript was always the
default option. This commit fixes that by remembering user's last
language used and creating a new block with that language.
We already had `rme-code-language` key to store users language
preference. We were also checking and using the store value from this
key if it existed. But we were not updating the storage to update this
key when user changes the language. This commit fixes that.
Fixes #6040.