Comment: fix hint incorrectly appearing if comment created from editor
## Reproduce
1. Open the comments sidebar, and ensure enough comments so that it is scrollable.
2. Highlight a text in the Editor and create a comment. The "↓" hint incorrectly appears on top of the new pending comment.

## Why
A comment created this way is counted immediately (in contrast with NewCommentForm), which then triggers the useEffect. Because `scrollToView` was in the comment, it will not necessarily scroll all the way to the bottom, thus our logic considers this as "!isAtBottom" and displays the hint.

## Solution
Add some buffer pixels in the consideration of `isAtBottom`. It's mostly whitespaces anyway.

## Remarks
This solution is not perfect because zoom levels will impact the actual amount. Hopefully the current value is sufficient.

The alternative is to use `threads.filter((c) => !c.isNew).length` to exclude the "new" comment, but it seems like a potential performance problem.