Editor: propagate extension changes while being mounted
## Scenario
- Open two tabs: (1) document (2) Preferences page
- Toggle "Smart Text" setting.

Notice that it has no effect despite the correct (updated) extensions array is passed in.

## Problem
`app/editor` basically has to re-initialize everything in the render. Tried to make a partial `this.init()` that doesn't end up creating multiple views, but that did not work as well.

## Approach
Just remount the entire component.

This will of course blow away transient states, but looking at the existing 4 states, they are either re-initialized or it doesn't matter if they are blown away.

One downside is the future -- if a new state is added, it must be ok with this re-mounting.

## Remark
This whole scenario is arguably ignorable -- it's a rare one, and users can just refresh to make it take effect.