Second attempt
Given that we are ok with grabbing everything on the js side (previously, I felt this defeated the purpose of pagination), we can use that call to circumvent the `offset/limit` with Sequelize -- by doing revocation-check on one call, and pagination on the other. This should also address the "return less than requested" problem since it is "pre-filtered".

In the second call, I changed to "all memberships", assuming that it's faster without having to check for userId again.