Restore highlight effect through URL params
Assuming the suggested Document.tsx method meant adding an @observable there, I think that would incur some prop-drilling of the setter since the navigator will be 2 levels down.

Since we want to avoid moving states to the store, I used URLParams to indicate which item to highlight. RevisionViewer just listens and updates accordingly.

For simplicity (avoiding the need to sync variables), the param is cleared on F5.